refactor #161 Hub 모듈 공통 응답 클래스 및 success 메서드 작성
-[] Hub 모듈의 공통 응답 클래스 ApiResponse에 T data 필드 추가
-[] 규격서의 응답 형식에 맞추기 위해 data 필드에 @JsonUnwrapped 적용
-[] CarInfoController의 모든 API에 공통 응답 형식 적용
-[] dto 패키지 위치를 application 계층으로 변경
