+++ tracky-consumer/src/main/resources/application.yml
@@ -1,9 +1,6 @@
 spring:
   config:
     import: classpath:application-common.yml
-  jpa:
-    hibernate:
-      ddl-auto: update
   application:
     name: tracky-consumer
