+++ tracky-hub/src/main/java/kernel360trackybe/trackyhub/application/dto/request/CycleGpsRequest.java
@@ -1,4 +1,4 @@
-package kernel360trackybe.trackyhub.presentation.dto;
+package kernel360trackybe.trackyhub.application.dto.request;
 
 import kernel360.trackycore.core.common.entity.vo.GpsInfo;
 import lombok.Getter;
