+++ tracky-hub/src/main/java/kernel360trackybe/trackyhub/application/dto/request/TokenRequest.java
@@ -1,4 +1,4 @@
-package kernel360trackybe.trackyhub.presentation.dto;
+package kernel360trackybe.trackyhub.application.dto.request;
 
 import kernel360.trackycore.core.common.entity.vo.EmulatorInfo;
 import lombok.Getter;
