+++ tracky-hub/src/main/java/kernel360trackybe/trackyhub/application/dto/response/TokenResponse.java
@@ -1,0 +1,12 @@
+package kernel360trackybe.trackyhub.application.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class TokenResponse {
+	private String mdn;
+	private String token;
+	private String exPeriod;
+}
