+++ tracky-hub/src/main/java/kernel360trackybe/trackyhub/presentation/CarInfoController.java
@@ -6,12 +6,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import kernel360trackybe.trackyhub.application.dto.request.CarOnOffRequest;
+import kernel360trackybe.trackyhub.application.dto.request.CycleInfoRequest;
+import kernel360trackybe.trackyhub.application.dto.request.TokenRequest;
+import kernel360trackybe.trackyhub.application.dto.response.ApiResponse;
+import kernel360trackybe.trackyhub.application.dto.response.TokenResponse;
 import kernel360trackybe.trackyhub.application.service.CarInfoProducerService;
-import kernel360trackybe.trackyhub.presentation.dto.ApiResponse;
-import kernel360trackybe.trackyhub.presentation.dto.ApiTokenResponse;
-import kernel360trackybe.trackyhub.presentation.dto.CarOnOffRequest;
-import kernel360trackybe.trackyhub.presentation.dto.CycleInfoRequest;
-import kernel360trackybe.trackyhub.presentation.dto.TokenRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -27,27 +27,28 @@ public class CarInfoController {
 	public ApiResponse sendCycleInfo(@RequestBody CycleInfoRequest cycleInfoRequest) {
 
 		producerService.sendCycleInfo(cycleInfoRequest);
-		return new ApiResponse("000", "Success", cycleInfoRequest.getMdn());
+		return ApiResponse.success(cycleInfoRequest.getMdn());
 	}
 
 	@PostMapping(value = "/on")
 	public ApiResponse sendCarStart(@RequestBody CarOnOffRequest carOnOffRequest) {
 
 		producerService.sendCarStart(carOnOffRequest);
-		return new ApiResponse("000", "Success", carOnOffRequest.getMdn());
+		return ApiResponse.success(carOnOffRequest.getMdn());
 	}
 
 	@PostMapping(value = "/off")
-	public ApiResponse sendCarStop(@RequestBody CarOnOffRequest carOnOffRequest) {
+	public ApiResponse<String> sendCarStop(@RequestBody CarOnOffRequest carOnOffRequest) {
 
 		producerService.sendCarStop(carOnOffRequest);
-		return new ApiResponse("000", "Success", carOnOffRequest.getMdn());
+		return ApiResponse.success(carOnOffRequest.getMdn());
 	}
 
 	@PostMapping(value = "/token")
-	public ApiTokenResponse getToken(@RequestBody TokenRequest tokenRequest) {
+	public ApiResponse<TokenResponse> getToken(@RequestBody TokenRequest tokenRequest) {
 
 		String token = producerService.getToken();
-		return new ApiTokenResponse("000", "Success", tokenRequest.getMdn(), token, "4");
+		TokenResponse tokenResponse = new TokenResponse(tokenRequest.getMdn(), token, "4");
+		return ApiResponse.success(tokenResponse);
 	}
 }
