--- tracky-hub/src/main/java/kernel360trackybe/trackyhub/presentation/dto/ApiTokenResponse.java
@@ -1,16 +1,0 @@
-package kernel360trackybe.trackyhub.presentation.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@AllArgsConstructor
-public class ApiTokenResponse {
-	private String rstCd;
-	private String rstMsg;
-	private String mdn;
-	private String token;
-	private String exPeriod;
-}
